Issue_#3:Add View page for a find license plate in the video
Improve design in for page with photos 
Improve design in for page with Video
