+++ LPRAutomatic/App.xaml
@@ -2,8 +2,8 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:LPRAutomatic"
-             StartupUri="ViewModel/MainWindow.xaml">
+             StartupUri="ViewModel/MainWindow.xaml"
+             Exit="Application_Exit">
     <Application.Resources>
-      
     </Application.Resources>
 </Application>
