+++ LPRAutomatic/Helper/MemoryDictionaryHelper.cs
@@ -1,11 +1,16 @@
 ﻿using LPRAutomatic.Model;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace LPRAutomatic.Helper
 {
     public static class MemoryDictionaryHelper
     {
-        private static IDictionary<string, LicensePlateModel> _infoLicensePlateModels = new Dictionary<string, LicensePlateModel>();
+        private static Dictionary<string, LicensePlateModel> _infoLicensePlateModels = new Dictionary<string, LicensePlateModel>();
+
+        public static int LastIndex = 0;
+
+        public static int AddedLastIndex { get { return _infoLicensePlateModels.Count; } }
 
         public static LicensePlateModel GetLicensePlate(string name)
         {
@@ -19,6 +24,16 @@ namespace LPRAutomatic.Helper
         {
             if (!string.IsNullOrEmpty(licensePlate.LicensePlate))
                 _infoLicensePlateModels[licensePlate.LicensePlate] = licensePlate;
+        }
+
+        public static int Count()
+        {
+            return _infoLicensePlateModels.Values.Count;
+        }
+
+        public static LicensePlateModel GetLatLicensePlate()
+        {
+            return _infoLicensePlateModels.Values.Last();
         }
     }
 }
