+++ LPRAutomatic/Helper/VideoCameraHepler.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace LPRAutomatic.Helper
+{
+    public static class VideoCameraHepler
+    {
+        public static bool CameraIsActive = true;
+    }
+}
