+++ LPRAutomatic/ViewModel/MainWindow.xaml
@@ -7,73 +7,26 @@
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800" Icon="/LPRAutomatic;component/LCPicon.ico" Background="#FF0C369F">
 
-    <Grid x:Name="MainGrid" Background="#FFCDCFD4">
-        <Grid x:Name="FileInfoGrid" Height="46" Margin="200,20,-0.4,0" VerticalAlignment="Top" Background="#FFCDCFD4" OpacityMask="#FFCDCFD4">
-            <TextBox x:Name="FileNameTextBox" Height="30" Margin="10,10,113,0" Text="File name" VerticalAlignment="Top" FontSize="14" IsUndoEnabled="True" TextOptions.TextFormattingMode="Display" BorderBrush="White"/>
-            <Button x:Name="FindButton" Content="Load Image" Margin="0,10,10,0" VerticalAlignment="Top" Height="30" Click="FindButton_Click" HorizontalAlignment="Right" Width="99" Background="#FF16A085" FontSize="14" TextOptions.TextFormattingMode="Display" Foreground="{DynamicResource {x:Static SystemColors.ControlBrushKey}}" BorderBrush="#FF16A085"/>
-        </Grid>
-        <Grid x:Name="InfoGrid" HorizontalAlignment="Left" Width="200" Background="#FF535E64" Margin="0,20,0,0">
-            <Grid.OpacityMask>
-                <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
-                    <GradientStop Color="Black" Offset="0"/>
-                    <GradientStop Color="#FF535E64" Offset="1"/>
-                </LinearGradientBrush>
-            </Grid.OpacityMask>
-            <ListBox x:Name="InfoNumberListBox" HorizontalAlignment="Left" Width="197" SelectionChanged="InfoNumberListBox_SelectionChanged" Background="#FF535E64" IsSynchronizedWithCurrentItem="True" Foreground="White" Margin="1,14,0,-1" BorderBrush="#FF535E64" RenderTransformOrigin="0.5,0.5">
-                <ListBox.RenderTransform>
-                    <TransformGroup>
-                        <ScaleTransform/>
-                        <SkewTransform/>
-                        <RotateTransform Angle="0.098"/>
-                        <TranslateTransform/>
-                    </TransformGroup>
-                </ListBox.RenderTransform>
-            </ListBox>
-        </Grid>
-        <Grid x:Name="CarImageGrid" Margin="210,66.133,-0.4,105" Background="#FFCDCFD4" Grid.RowSpan="2" Grid.Row="1">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="27*"/>
-                <RowDefinition Height="223*"/>
-            </Grid.RowDefinitions>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition/>
-                <ColumnDefinition Width="568*"/>
-            </Grid.ColumnDefinitions>
-            <Image x:Name="CarImage" Margin="9.2,10,10,10" OpacityMask="{DynamicResource {x:Static SystemColors.ControlLightLightBrushKey}}" Grid.Column="1" Grid.RowSpan="2"/>
-        </Grid>
-        <Grid x:Name="NumberImagesGrid" Margin="210,0,-0.4,0" Height="115" VerticalAlignment="Bottom" Background="#FFCDCFD4" Grid.Row="2">
-            <Image x:Name="ImageOriginal" Margin="0,10,10,10" HorizontalAlignment="Right" Width="335" RenderTransformOrigin="0.5,0.5">
-                <Image.RenderTransform>
-                    <TransformGroup>
-                        <ScaleTransform/>
-                        <SkewTransform/>
-                        <RotateTransform Angle="-0.203"/>
-                        <TranslateTransform/>
-                    </TransformGroup>
-                </Image.RenderTransform>
-            </Image>
-            <Grid x:Name="InfoOperationGrid" HorizontalAlignment="Left" Height="95" Margin="10,10,0,0" VerticalAlignment="Top" Width="218">
-                <Label x:Name="TimaLabel" HorizontalAlignment="Right" Margin="0,0,10,59" VerticalAlignment="Bottom" Width="123" Height="29" FontSize="14"/>
-                <Label x:Name="GuantityPlateLable" Content="" HorizontalAlignment="Right" Margin="0,59,19,0" VerticalAlignment="Top" Width="70"/>
-                <Label Content="Time:" HorizontalAlignment="Left" Margin="10,6,0,0" VerticalAlignment="Top" Width="70" FontSize="14" Height="31" Foreground="#FF428279"/>
-                <Label Content="Count result:" HorizontalAlignment="Left" Margin="10,59,0,0" VerticalAlignment="Top" Width="114" FontSize="14" Foreground="#FF428279"/>
-            </Grid>
+    <Grid x:Name="MainGrid" Background="White">
 
-        </Grid>
         <Grid Height="20" VerticalAlignment="Top" Background="#FF303D45" Margin="0,0,-0.4,0">
             <Menu  Height="Auto" Width="Auto" VerticalAlignment="Bottom" Background="#FF303D45" HorizontalAlignment="Left" HorizontalContentAlignment="Stretch" VerticalContentAlignment="Stretch" >
                 <MenuItem x:Name="FileMenuItem" Width="99" Height="20" Header="File" Foreground="#FFCDCFD4" Background="#FF303D45">
-                    <MenuItem Header="Photo" Foreground="#FFCDCFD4" Background="#FF303D45" ForceCursor="False" />
-                    <MenuItem Header="Video" Foreground="#FFCDCFD4" Background="#FF303D45" Click="MenuItem_Click"/>
-                    <MenuItem Header="Celar all list" Foreground="#FFCDCFD4" Background="#FF303D45"/>
+                    <MenuItem x:Name="PhotoMenuItem" Header="Photo" Foreground="#FFCDCFD4" Background="#FF303D45" ForceCursor="False" Click="PhotoMenuItem_Click"/>
+                    <MenuItem x:Name="VideoMenuItem" Header="Video" Foreground="#FFCDCFD4" Background="#FF303D45" Click="VideoMenuItem_Click"/>
                 </MenuItem>
-                <MenuItem x:Name="SettingMenuItem" Width="99" Height="20" Header="Tolls" Foreground="#FFCDCFD4"/>
-                <MenuItem x:Name="InfoMenuItem" Width="99" Height="20" Header="Help" Foreground="#FFCDCFD4"/>
+                <MenuItem x:Name="SettingMenuItem" Width="99" Height="20" Header="Tolls" Foreground="#FFCDCFD4">
+                    <MenuItem x:Name="SaveToExcel" Header="Save to Excel file" Foreground="#FFCDCFD4" Background="#FF303D45"/>
+                </MenuItem>
+                <MenuItem x:Name="InfoMenuItem" Width="99" Height="20" Header="Help" Foreground="#FFCDCFD4">
+                    <MenuItem x:Name="ViewHelp" Header="View Help" Foreground="#FFCDCFD4" Background="#FF303D45"/>
+                    <MenuItem x:Name="About" Header="About" Foreground="#FFCDCFD4" Background="#FF303D45"/>
+                </MenuItem> >
 
             </Menu>
         </Grid>
-        <!--<Frame HorizontalAlignment="Stretch" VerticalAlignment="Stretch" HorizontalContentAlignment="Stretch" VerticalContentAlignment="Stretch" x:Name="NavigationFrame" NavigationUIVisibility="Hidden" Margin="0,20,-0.4,0"/>-->
-        <Frame x:Name="NavigationFrame" Grid.RowSpan="2" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" HorizontalContentAlignment="Stretch" VerticalContentAlignment="Stretch" NavigationUIVisibility="Hidden" Margin="0, 20, 0, 0"/>
+        <Frame x:Name="NavigationFrame" Grid.RowSpan="2" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" HorizontalContentAlignment="Stretch" VerticalContentAlignment="Stretch" NavigationUIVisibility="Hidden" Margin="0, 20, 0, 0" Panel.ZIndex="1"/>
+        <Image x:Name="MainImage" HorizontalAlignment="Stretch" Margin="60,80,60,60" VerticalAlignment="Stretch" OpacityMask="#FF3C3535"/>
 
     </Grid>
 </Window>
