+++ LPRAutomatic/ViewModel/PhotoLPRWindow.xaml
@@ -1,0 +1,63 @@
+﻿<Page x:Class="LPRAutomatic.ViewModel.PhotoLPRWindow"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:LPRAutomatic.ViewModel"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="PhotoLPRWindow">
+    <Grid>
+        <Grid x:Name="FileInfoGrid" Height="46" Margin="200,0,0,0" VerticalAlignment="Top" Background="#FFCDCFD4" OpacityMask="#FFCDCFD4">
+            <TextBox x:Name="FileNameTextBox" Height="30" Margin="10,10,113,0" Text="File name" VerticalAlignment="Top" FontSize="14" IsUndoEnabled="True" TextOptions.TextFormattingMode="Display" BorderBrush="White"/>
+            <Button x:Name="FindButton" Content="Load Image" Margin="0,10,10,0" VerticalAlignment="Top" Height="30" Click="FindButton_Click" HorizontalAlignment="Right" Width="99" Background="#FF16A085" FontSize="14" TextOptions.TextFormattingMode="Display" Foreground="{DynamicResource {x:Static SystemColors.ControlBrushKey}}" BorderBrush="#FF16A085"/>
+        </Grid>
+        <Grid x:Name="InfoGrid" HorizontalAlignment="Left" Width="200" Background="#FF535E64" Margin="0,0,0,-0.4">
+            <Grid.OpacityMask>
+                <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
+                    <GradientStop Color="Black" Offset="0"/>
+                    <GradientStop Color="#FF535E64" Offset="1"/>
+                </LinearGradientBrush>
+            </Grid.OpacityMask>
+            <ListBox x:Name="InfoNumberListBox" HorizontalAlignment="Left" Width="197" SelectionChanged="InfoNumberListBox_SelectionChanged" Background="#FF535E64" IsSynchronizedWithCurrentItem="True" Foreground="White" Margin="1,2,0,-1" BorderBrush="#FF535E64" RenderTransformOrigin="0.5,0.5" FontSize="14">
+                <ListBox.RenderTransform>
+                    <TransformGroup>
+                        <ScaleTransform/>
+                        <SkewTransform/>
+                        <RotateTransform Angle="0.098"/>
+                        <TranslateTransform/>
+                    </TransformGroup>
+                </ListBox.RenderTransform>
+            </ListBox>
+        </Grid>
+        <Grid x:Name="CarImageGrid" Margin="200,46,0,104.6" Background="#FFCDCFD4">
+            <Grid.RowDefinitions>
+                <RowDefinition Height="8*"/>
+                <RowDefinition Height="85*"/>
+            </Grid.RowDefinitions>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition/>
+                <ColumnDefinition Width="568*"/>
+            </Grid.ColumnDefinitions>
+            <Image x:Name="CarImage" Margin="9.2,10,10,10.2" OpacityMask="{DynamicResource {x:Static SystemColors.ControlLightLightBrushKey}}" Grid.Column="1" Grid.RowSpan="2"/>
+        </Grid>
+        <Grid x:Name="NumberImagesGrid" Margin="200,0,0,-0.4" Height="115" VerticalAlignment="Bottom" Background="#FFCDCFD4">
+            <Image x:Name="ImageOriginal" Margin="0,10,10,10" HorizontalAlignment="Right" Width="335" RenderTransformOrigin="0.5,0.5">
+                <Image.RenderTransform>
+                    <TransformGroup>
+                        <ScaleTransform/>
+                        <SkewTransform/>
+                        <RotateTransform Angle="-0.203"/>
+                        <TranslateTransform/>
+                    </TransformGroup>
+                </Image.RenderTransform>
+            </Image>
+            <Grid x:Name="InfoOperationGrid" HorizontalAlignment="Left" Height="95" Margin="10,10,0,0" VerticalAlignment="Top" Width="218">
+                <Label x:Name="TimaLabel" HorizontalAlignment="Right" Margin="0,0,10,59" VerticalAlignment="Bottom" Width="123" Height="29" FontSize="14"/>
+                <Label x:Name="GuantityPlateLable" Content="" HorizontalAlignment="Right" Margin="0,59,19,0" VerticalAlignment="Top" Width="70"/>
+                <Label Content="Time:" HorizontalAlignment="Left" Margin="10,6,0,0" VerticalAlignment="Top" Width="70" FontSize="14" Height="31" Foreground="#FF428279"/>
+                <Label Content="Count result:" HorizontalAlignment="Left" Margin="10,59,0,0" VerticalAlignment="Top" Width="114" FontSize="14" Foreground="#FF428279"/>
+            </Grid>
+        </Grid>
+    </Grid>
+</Page>
