+++ LPRAutomatic/ViewModel/VideoLPRWindow.xaml
@@ -11,12 +11,21 @@
     <Grid Background="#FFCDCFD4" Name="VideoMainGrid" Margin="0,0,0,-0.4">
         <Grid HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="190,0,0,118">
             <Image x:Name="VideoImage" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Stretch="Fill" OpacityMask="{DynamicResource {x:Static SystemColors.ControlLightLightBrushKey}}"  Margin="0,0,0,-3.4"/>
-
         </Grid>
         <Grid x:Name="InfoPlateGrid" HorizontalAlignment="Left" VerticalAlignment="Stretch" Background="#FF428079" Width="190">
-            <ListBox HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Background="#FF428079"/>
+            <ListBox x:Name="LicensePlateList" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Background="#FF428079" Foreground="White" FontSize="14">
+                <ListBox.RenderTransform>
+                    <TransformGroup>
+                        <ScaleTransform/>
+                        <SkewTransform/>
+                        <RotateTransform Angle="0.098"/>
+                        <TranslateTransform/>
+                    </TransformGroup>
+                </ListBox.RenderTransform>
+            </ListBox>
         </Grid>
-        <Grid x:Name="DetailinformationGrid" HorizontalAlignment="Stretch" VerticalAlignment="Stretch" Margin="190,332,0,0"/>
-
+        <Grid x:Name="DetailinformationGrid" HorizontalAlignment="Left" VerticalAlignment="Bottom" Margin="190,0,0,0" Width="610" Height="118">
+            <Image x:Name="RecognizedPlateImage" Margin="10,10,300,10" VerticalAlignment="Stretch" HorizontalAlignment="Stretch"/>
+        </Grid>
     </Grid>
 </Page>
